requirements.txt
# Core dependencies
requests>=2.28.0
PyYAML>=6.0
python-dotenv>=1.0.0
psutil>=5.9.0

# CLI
click>=8.1.0

# Scheduling and calendar
icalendar>=5.0.0
pytz>=2023.3

# Monitoring and metrics  
prometheus-client>=0.17.0
aiohttp>=3.8.0

# Optional: Discord notifications
# discord.py>=2.3.0

# Optional: Telegram notifications
# python-telegram-bot>=20.0